requirements.txt
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
llama-index-core
llama-index-embeddings-langchain
llama-index-llms-anyscale
llama-index-llms-fireworks
llama-index-llms-langchain
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-readers-s3
llama-index-readers-file
llama-index-vector-stores-mongodb==0.1.6
streamlit
fastapi
uvicorn
pymongo